refactor: data 디렉터리 생성
